Add warning when we cannot find procdump.
