+++ src/Microsoft.TestPlatform.Extensions.BlameDataCollector/ICrashDumper.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using System.Collections.Generic;
 
 #nullable disable
@@ -9,7 +10,7 @@ namespace Microsoft.TestPlatform.Extensions.BlameDataCollector;
 
 public interface ICrashDumper
 {
-    void AttachToTargetProcess(int processId, string outputDirectory, DumpTypeOption dumpType, bool collectAlways);
+    void AttachToTargetProcess(int processId, string outputDirectory, DumpTypeOption dumpType, bool collectAlways, Action<string> logWarning);
 
     void WaitForDumpToFinish();
