+++ src/Microsoft.TestPlatform.Extensions.BlameDataCollector/Interfaces/IProcessDumpUtility.cs
@@ -38,7 +38,7 @@ public interface IProcessDumpUtility
     /// <param name="collectAlways">
     /// Collect the dump on process exit even if there is no exception
     /// </param>
-    void StartTriggerBasedProcessDump(int processId, string testResultsDirectory, bool isFullDump, string targetFramework, bool collectAlways);
+    void StartTriggerBasedProcessDump(int processId, string testResultsDirectory, bool isFullDump, string targetFramework, bool collectAlways, Action<string> logWarning);
 
     /// <summary>
     /// Launch proc dump process to capture dump in case of a testhost hang and wait for it to exit
